Fixed changelog for rc00, some minor codefixes, reorganizations of code. Reverted some namechanges to make the changediff less of an eyesore
