+++ src/lib/SerializationUtils.ts
